+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [7.7.3] - 2023-03-21
+### Changed
+- do not include values that are zeros
+
 ## [7.7.2] - 2023-03-09
 ### Changed
 - update umbrella SDK